refactor: cleanup unnecessary functions
