+++ apps/admin-ui/src/component/applications/queries.ts
@@ -1,1 +1,1 @@
-export { APPLICATION_ADMIN_QUERY as APPLICATION_QUERY } from "common/src/queries/application";
+export { APPLICATION_ADMIN_QUERY } from "common/src/queries/application";
